Remove utils.rs and move shared logic to common.rs
